\name{SPARSEMODr-package}
\alias{SPARSEMODr-package}
\alias{SPARSEMODr}
\docType{package}
\title{
  SPAtial Resolution-SEnsitive Models of Outbreak Dynamics
}
\description{
  Implementation of spatially-explicit, stochastic disease models with customizable time windows that describe how parameter values fluctuate over the course of the simulation (e.g., in response to public health or conservation interventions).
}
\details{
    Stochastic realizations of the models can be performed in parallel using the \code{\link{model_parallel}} functions, and time-varying parameters can be supplied using \code{\link{time_windows}} objects.
}
\author{

Maintainer: Joseph Mihaljevic <joseph.mihaljevic@nau.edu>
}

\keyword{ package }
